improve structure of extendables
