+++ src/extendables/User/Minion.ts
@@ -1,14 +1,16 @@
 import { User } from 'discord.js';
-import { Extendable, ExtendableStore, KlasaUser } from 'klasa';
+import { Extendable, ExtendableStore, KlasaClient, KlasaUser, SettingsFolder } from 'klasa';
 import Monster from 'oldschooljs/dist/structures/Monster';
 
-import { Activity, Emoji, Time } from '../../lib/constants';
+import { production } from '../../config';
+import { Activity, Channel, Emoji, Events, MAX_QP, PerkTier, Time } from '../../lib/constants';
 import ClueTiers from '../../lib/minions/data/clueTiers';
 import killableMonsters from '../../lib/minions/data/killableMonsters';
 import { MinigameIDsEnum } from '../../lib/minions/data/minigames';
 import { Planks } from '../../lib/minions/data/planks';
 import { GroupMonsterActivityTaskOptions } from '../../lib/minions/types';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
+import Skills from '../../lib/skilling/skills';
 import Agility from '../../lib/skilling/skills/agility';
 import Cooking from '../../lib/skilling/skills/cooking';
 import Crafting from '../../lib/skilling/skills/crafting';
@@ -20,6 +22,7 @@ import Runecraft, { RunecraftActivityTaskOptions } from '../../lib/skilling/skil
 import Smithing from '../../lib/skilling/skills/smithing';
 import Woodcutting from '../../lib/skilling/skills/woodcutting';
 import { SkillsEnum } from '../../lib/skilling/types';
+import { ItemBank } from '../../lib/types';
 import {
 	AgilityActivityTaskOptions,
 	AlchingActivityTaskOptions,
@@ -40,8 +43,19 @@ import {
 	WoodcuttingActivityTaskOptions,
 	ZalcanoActivityTaskOptions
 } from '../../lib/types/minions';
-import { formatDuration, itemNameFromID } from '../../lib/util';
+import {
+	addBanks,
+	addItemToBank,
+	convertXPtoLVL,
+	formatDuration,
+	itemNameFromID,
+	toTitleCase,
+	Util
+} from '../../lib/util';
+import { channelIsSendable } from '../../lib/util/channelIsSendable';
+import { formatOrdinal } from '../../lib/util/formatOrdinal';
 import getActivityOfUser from '../../lib/util/getActivityOfUser';
+import getUsersPerkTier from '../../lib/util/getUsersPerkTier';
 import { NightmareActivityTaskOptions } from './../../lib/types/minions';
 
 export default class extends Extendable {
@@ -321,11 +335,203 @@ export default class extends Extendable {
 		return this.settings.get(UserSettings.MonsterScores)[monster.id] ?? 0;
 	}
 
-	getCL(this: KlasaUser, itemID: number) {
-		return this.settings.get(UserSettings.CollectionLogBank)[itemID] ?? 0;
-	}
-
 	getMinigameScore(this: KlasaUser, id: MinigameIDsEnum) {
 		return this.settings.get(UserSettings.MinigameScores)[id] ?? 0;
+	}
+
+	// @ts-ignore 2784
+	public get hasMinion(this: User) {
+		return this.settings.get(UserSettings.Minion.HasBought);
+	}
+
+	// @ts-ignore 2784
+	public get maxTripLength(this: User) {
+		const perkTier = getUsersPerkTier(this);
+		if (perkTier === PerkTier.Two) return Time.Minute * 33;
+		if (perkTier === PerkTier.Three) return Time.Minute * 36;
+		if (perkTier >= PerkTier.Four) return Time.Minute * 40;
+
+		return Time.Minute * 30;
+	}
+
+	// @ts-ignore 2784
+	public get minionIsBusy(this: User): boolean {
+		const usersTask = getActivityOfUser(this.client, this.id);
+		return Boolean(usersTask);
+	}
+
+	// @ts-ignore 2784
+	public get minionName(this: User): string {
+		const name = this.settings.get(UserSettings.Minion.Name);
+		const prefix = this.settings.get(UserSettings.Minion.Ironman) ? Emoji.Ironman : '';
+
+		const icon = this.settings.get(UserSettings.Minion.Icon) ?? Emoji.Minion;
+
+		return name
+			? `${prefix} ${icon} **${Util.escapeMarkdown(name)}**`
+			: `${prefix} ${icon} Your minion`;
+	}
+
+	public async incrementMinionDailyDuration(this: User, duration: number) {
+		await this.settings.sync(true);
+
+		const currentDuration = this.settings.get(UserSettings.Minion.DailyDuration);
+		const newDuration = currentDuration + duration;
+		if (newDuration > Time.Hour * 18) {
+			const log = `[MOU] Minion has been active for ${formatDuration(newDuration)}.`;
+
+			this.log(log);
+			if (production) {
+				const channel = this.client.channels.get(Channel.ErrorLogs);
+				if (channelIsSendable(channel)) {
+					channel.send(`${this.sanitizedName} ${log}`);
+				}
+			}
+		}
+
+		return this.settings.update(UserSettings.Minion.DailyDuration, newDuration);
+	}
+
+	public async addXP(this: User, skillName: SkillsEnum, amount: number) {
+		await this.settings.sync(true);
+		const currentXP = this.settings.get(`skills.${skillName}`) as number;
+		if (currentXP >= 200_000_000) return;
+
+		const skill = Object.values(Skills).find(skill => skill.id === skillName);
+		if (!skill) return;
+
+		const newXP = Math.min(200_000_000, currentXP + amount);
+
+		// If they reached a XP milestone, send a server notification.
+		for (const XPMilestone of [50_000_000, 100_000_000, 150_000_000, 200_000_000]) {
+			if (newXP < XPMilestone) break;
+
+			if (currentXP < XPMilestone && newXP >= XPMilestone) {
+				this.client.emit(
+					Events.ServerNotification,
+					`${skill.emoji} **${this.username}'s** minion, ${
+						this.minionName
+					}, just achieved ${newXP.toLocaleString()} XP in ${toTitleCase(skillName)}!`
+				);
+				break;
+			}
+		}
+
+		// If they just reached 99, send a server notification.
+		if (convertXPtoLVL(currentXP) < 99 && convertXPtoLVL(newXP) >= 99) {
+			const skillNameCased = toTitleCase(skillName);
+			const [usersWith] = await this.client.query<
+				{
+					count: string;
+				}[]
+			>(`SELECT COUNT(*) FROM users WHERE "skills.${skillName}" > 13034430;`);
+
+			this.client.emit(
+				Events.ServerNotification,
+				`${skill.emoji} **${this.username}'s** minion, ${
+					this.minionName
+				}, just achieved level 99 in ${skillNameCased}! They are the ${formatOrdinal(
+					parseInt(usersWith.count) + 1
+				)} to get 99 ${skillNameCased}.`
+			);
+		}
+
+		return this.settings.update(`skills.${skillName}`, Math.floor(newXP));
+	}
+
+	public skillLevel(this: User, skillName: SkillsEnum) {
+		return convertXPtoLVL(this.settings.get(`skills.${skillName}`) as number);
+	}
+
+	public totalLevel(this: User, returnXP = false) {
+		const userXPs = Object.values(
+			(this.settings.get('skills') as SettingsFolder).toJSON() as Record<string, number>
+		);
+		let totalLevel = 0;
+		for (const xp of userXPs) {
+			totalLevel += returnXP ? xp : convertXPtoLVL(xp);
+		}
+		return totalLevel;
+	}
+
+	// @ts-ignore 2784
+	get isBusy(this: User) {
+		const client = this.client as KlasaClient;
+		return (
+			client.oneCommandAtATimeCache.has(this.id) || client.secondaryUserBusyCache.has(this.id)
+		);
+	}
+
+	/**
+	 * Toggle whether this user is busy or not, this adds another layer of locking the user
+	 * from economy actions.
+	 *
+	 * @param busy boolean Whether the new toggled state will be busy or not busy.
+	 */
+	public toggleBusy(this: User, busy: boolean) {
+		const client = this.client as KlasaClient;
+
+		if (busy) {
+			client.secondaryUserBusyCache.add(this.id);
+		} else {
+			client.secondaryUserBusyCache.delete(this.id);
+		}
+	}
+
+	public async addQP(this: User, amount: number) {
+		await this.settings.sync(true);
+		const currentQP = this.settings.get(UserSettings.QP);
+		const newQP = Math.min(MAX_QP, currentQP + amount);
+
+		if (currentQP < MAX_QP && newQP === MAX_QP) {
+			this.client.emit(
+				Events.ServerNotification,
+				`${Emoji.QuestIcon} **${this.username}'s** minion, ${this.minionName}, just achieved the maximum amount of Quest Points!`
+			);
+		}
+
+		this.log(`had ${newQP} QP added. Before[${currentQP}] New[${newQP}]`);
+		return this.settings.update(UserSettings.QP, newQP);
+	}
+
+	// @ts-ignore 2784
+	public get isIronman(this: User) {
+		return this.settings.get(UserSettings.Minion.Ironman);
+	}
+
+	public async incrementMonsterScore(this: User, monsterID: number, amountToAdd = 1) {
+		await this.settings.sync(true);
+		const currentMonsterScores = this.settings.get(UserSettings.MonsterScores);
+
+		this.log(`had Quantity[${amountToAdd}] KC added to Monster[${monsterID}]`);
+
+		return this.settings.update(
+			UserSettings.MonsterScores,
+			addItemToBank(currentMonsterScores, monsterID, amountToAdd)
+		);
+	}
+
+	public async incrementClueScore(this: User, clueID: number, amountToAdd = 1) {
+		await this.settings.sync(true);
+		const currentClueScores = this.settings.get(UserSettings.ClueScores);
+
+		this.log(`had Quantity[${amountToAdd}] KC added to Clue[${clueID}]`);
+
+		return this.settings.update(
+			UserSettings.ClueScores,
+			addItemToBank(currentClueScores, clueID, amountToAdd)
+		);
+	}
+
+	public async incrementMinigameScore(this: User, minigameID: number, amountToAdd = 1) {
+		await this.settings.sync(true);
+		const currentMinigameScores = this.settings.get(UserSettings.MinigameScores);
+
+		this.log(`had Quantity[${amountToAdd}] Score added to Minigame[${minigameID}]`);
+
+		return this.settings.update(
+			UserSettings.MinigameScores,
+			addItemToBank(currentMinigameScores, minigameID, amountToAdd)
+		);
 	}
 }
